layout mismatch cleanup and some comments
